Module rsp2_tasks::entry_points

Structs

AppendLog
Used by entry points that have no trial directory, but where the user may wish to save log output to a specified path, at their choosing.

AppendLogInner

OptionalFileType

Functions

after_diagonalization

bond_test

compute_for_phonopy

converge_vdw

dynmat_analysis

dynmat_at_q

layer_mode_frequencies

make_supercell

plot_vdw

print_library_paths

rerun_analysis

rsp2

rsp2_acgsd

rsp2_acgsd_and_dynmat

shear_plot

sparse_analysis

test_rayon