Struct rsp2_tasks::exposed_for_testing::ForceConstants

pub struct ForceConstants(pub RawCsr<M3<V3<f64>>, PrimI, SuperI>);

Force constants matrix, with only rows for atoms in ForceConstants::DESIGNATED_CELL stored.

Implementations

impl ForceConstants

pub const DESIGNATED_CELL: [u32; 3]

Displaced atoms must always be in this cell.

pub fn compute_required_rows(
super_displacements: &[(usize, V3<f64>)],
force_sets: &[BTreeMap<usize, V3<f64>>],
cart_rots: &[M3<V3<f64>>],
super_deperms: &[Perm],
sc: &SupercellToken
) -> Result<ForceConstants, Error>

Compute a partially filled force constants matrix that contains only the rows that appear in the definition of the dynamical matrix. (those where the displaced atom is an image in ForceConstants::DESIGNATED_CELL)

This basically follows Phonopy’s method for --writefc to a T, although it is a bit smarter in some places, and probably dumber in other places.

Notable differences from phonopy include:

This uses a sparse format for anything indexed by atoms (primitive or super)

Phonopy translates each displaced atom to the origin and performs expensive brute force overlap searches, while this leaves everything where they are and adjusts indices to simulate translation by a lattice point)

impl ForceConstants

pub fn dynmat_at_cart_q(
&self,
super_coords: &Coords,
qpoint_cart: V3<f64>,
sc: &SupercellToken,
masses: &[f64]
) -> DynamicalMatrix

Compute the dynamical matrix at a q-point.

The force constants do not need to contain data for rows outside the designated cell. (but if they do, it won’t hurt)

impl ForceConstants

pub fn to_super_force_constants_with_zeroed_rows(
&self,
sc: &SupercellToken
) -> SuperForceConstants

Convert to a type capable of containing data in arbitrary supercell rows, but do not yet actually fill the rows outside ForceConstants::DESIGNATED_CELL. (the data will still be sparse; this mostly just replaces primitive indices with supercell indices)

Why? Well, it provides a Permute impl, which could be used to construct something closer to the output of another tool like phonopy. It also provides conversions to and from dense matrices, for debugging and testing.

pub fn to_super_force_constants_with_all_rows(
&self,
sc: &SupercellToken
) -> SuperForceConstants

impl ForceConstants

pub fn impose_translational_invariance(
self,
sc: &SupercellToken,
level: u32
) -> ForceConstants

Imposes translational invariance like Phonopy.

Warning: This causes the ForceConstants matrix to effectively become dense!