Module rsp2_tasks::meta

Site metadata

High-level code carts site metadata around in the form of HLists of Rc<[T]> (or similar, like Option<Rc<[T]>>). The use of Rc eliminates concern about ownership (it is distinctly un-fun to turn an Option<Vec<T>> into an Option<&[T]), and allows this module to get away with a very small number of extension traits. (where a reasonably ergonomic API that can handle differences in ownership would require at least half a dozen)

Modules

prelude

Structs

Element
Represents a specific atomic number.

Layer

Mass

Traits

MetaPick
Get a single item of a metadata list by type.

MetaSendable
Workaround to use metadata where thread safety is required.

MetaSift
Get a subset of a metadata list by type, in any order.

Type Definitions

CartBonds

FracBonds

LayerScMatrices

SiteElements

SiteLayers

SiteMasses