Trait rsp2_tasks::meta::MetaSift

pub trait MetaSift<Targets, Indices> {
    fn sift(&self) -> Targets;
}

Get a subset of a metadata list by type, in any order.

Equivalent to list.clone().sculpt().0.

Intended for use on Metadata lists, which are (relatively) cheap to clone.

Required methods

fn sift(&self) -> Targets

Loading content...